feat(storage) split storage volume form in dedicated component to make it resusable for editing
