+++ src/pages/storage/forms/StorageVolumeCreate.tsx
@@ -1,0 +1,107 @@
+import React, { FC, useState } from "react";
+import { Button, Col, Row, useNotify } from "@canonical/react-components";
+import { useFormik } from "formik";
+import * as Yup from "yup";
+import { useQueryClient } from "@tanstack/react-query";
+import { queryKeys } from "util/queryKeys";
+import SubmitButton from "components/SubmitButton";
+import { createStorageVolume } from "api/storage-pools";
+import NotificationRow from "components/NotificationRow";
+import { useNavigate, useParams } from "react-router-dom";
+import { testDuplicateName } from "util/storageVolume";
+import BaseLayout from "components/BaseLayout";
+import {
+  StorageVolumeFormValues,
+  volumeFormToPayload,
+} from "pages/storage/forms/StorageVolumeForm";
+import StorageVolumeForm from "pages/storage/forms/StorageVolumeForm";
+
+const StorageVolumeCreate: FC = () => {
+  const navigate = useNavigate();
+  const notify = useNotify();
+  const queryClient = useQueryClient();
+  const controllerState = useState<AbortController | null>(null);
+  const { project, pool } = useParams<{ project: string; pool: string }>();
+
+  if (!project) {
+    return <>Missing project</>;
+  }
+
+  if (!pool) {
+    return <>Missing storage pool name</>;
+  }
+
+  const StorageSchema = Yup.object().shape({
+    name: Yup.string()
+      .test(...testDuplicateName(project, pool, controllerState))
+      .required("This field is required"),
+  });
+
+  const formik = useFormik<StorageVolumeFormValues>({
+    initialValues: {
+      content_type: "filesystem",
+      name: "",
+      project: project,
+      pool: pool,
+      size: "",
+      isReadOnly: false,
+      isCreating: true,
+    },
+    validationSchema: StorageSchema,
+    onSubmit: (values) => {
+      const volume = volumeFormToPayload(values, project);
+      createStorageVolume(pool, project, volume)
+        .then(() => {
+          void queryClient.invalidateQueries({
+            queryKey: [queryKeys.storage],
+          });
+          navigate(
+            `/ui/project/${project}/storage/detail/${pool}/volumes`,
+            notify.queue(
+              notify.success(`Storage volume ${values.name} created.`)
+            )
+          );
+        })
+        .catch((e) => {
+          formik.setSubmitting(false);
+          notify.failure("Storage volume creation failed", e);
+        });
+    },
+  });
+
+  const submitForm = () => {
+    void formik.submitForm();
+  };
+
+  return (
+    <BaseLayout title="Create volume" contentClassName="storage-volume-form">
+      <NotificationRow />
+      <StorageVolumeForm formik={formik} />
+      <div className="l-footer--sticky p-bottom-controls">
+        <hr />
+        <Row className="u-align--right">
+          <Col size={12}>
+            <Button
+              appearance="base"
+              onClick={() =>
+                navigate(
+                  `/ui/project/${project}/storage/detail/${pool}/volumes`
+                )
+              }
+            >
+              Cancel
+            </Button>
+            <SubmitButton
+              isSubmitting={formik.isSubmitting}
+              isDisabled={!formik.isValid}
+              onClick={submitForm}
+              buttonLabel="Create"
+            />
+          </Col>
+        </Row>
+      </div>
+    </BaseLayout>
+  );
+};
+
+export default StorageVolumeCreate;
